requirements.txt
gradio
sentence-transformers
transformers
faiss-cpu
numpy
PyPDF2
docx2txt
